moved storeUserData function into ApplicantPresenter, data is now saved to database
