+++ frontend/package.json
@@ -2,7 +2,7 @@
   "name": "IV1201 project client",
   "version": "0.1.0",
   "private": true,
-  "homepage": "/login",
+  "homepage": "/",
   "dependencies": {
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
